pytest: Separate webview tests
